fix(symthaea): Round 24 — platform guard, Phi precision, resonator dedup, feedback proptests
- Add compile-time 64-bit platform assertion (usize overflow guard)
- Remove lossy f64→f32→f64 round-trip in consciousness optimizer Phi delta
- Eliminate double cosine similarity computation in resonator recall
  (compute once, reuse for both max-sim and argmax)
- Add 3 property tests for feedback consensus edge cases: empty channel
  returns cycle-start, uniform proposals produce expected sum, oscillating
  feedback remains bounded over 100 cycles

+++ symthaea/symthaea-core/src/hdc/consciousness_optimizer.rs
@@ -148,7 +148,7 @@ impl ConsciousnessOptimizer {
             if current_phi > prev_phi {
                 let action_vec = BinaryHV::random((5000 + action_taken) as u64);
                 let high_phi_vec = BinaryHV::random(6000); // Represents high consciousness
-                let strength = ((current_phi - prev_phi) as f32).min(1.0) as f64;
+                let strength = (current_phi - prev_phi).min(1.0);
 
                 self.causal_model
                     .add_causal_link(action_vec, high_phi_vec, strength);
